Refactor theme management by consolidating theme switching logic into app.js and removing the obsolete theme-switcher.js file. Updated app layout to include the default theme attribute in the HTML tag for improved theme handling.
